Exit if dockerized postgres is already running
This creates the postgres container with a "postgres_<timestamp>" name.
This lets us check for an existing, running, postgres container and
exit early if it already exists. It's a little more clear/explicit than
the "port 5432 already in use" error message that results from rerunning
the script, and it addresses the gotcha in footnote 30